Fixed project dtuff
