+++ .gitignore
@@ -1,0 +1,3 @@
+/Math-Madness/nbproject
+/Math-Madness/build.xml
+/MathMadness/manifest.mf
