--- JavaFXApplication10/nbproject/project.xml
@@ -1,25 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://www.netbeans.org/ns/project/1">
-    <type>org.netbeans.modules.java.j2seproject</type>
-    <configuration>
-        <buildExtensions xmlns="http://www.netbeans.org/ns/ant-build-extender/1">
-            <extension file="jfx-impl.xml" id="jfx3">
-                <dependency dependsOn="-jfx-copylibs" target="-post-jar"/>
-                <dependency dependsOn="-rebase-libs" target="-post-jar"/>
-                <dependency dependsOn="jfx-deployment" target="-post-jar"/>
-                <dependency dependsOn="jar" target="debug"/>
-                <dependency dependsOn="jar" target="profile"/>
-                <dependency dependsOn="jar" target="run"/>
-            </extension>
-        </buildExtensions>
-        <data xmlns="http://www.netbeans.org/ns/j2se-project/3">
-            <name>JavaFXApplication10</name>
-            <source-roots>
-                <root id="src.dir"/>
-            </source-roots>
-            <test-roots>
-                <root id="test.src.dir"/>
-            </test-roots>
-        </data>
-    </configuration>
-</project>
